Renamed TCM::Report::Class to Report::Instance
I also changed the relevant methods in that class, as well as other classes
that use this class somehow.

As part of this, I renamed various $report_class vars & report_class
attributes to instance_report. The former naming is very confusing, since it
can be read as "the class to use for reports" as opposed to "a report on a
class", whereas "instance_report" is clearly "a report on an instance".

I didn't add any back compat bits since the reporting code is clearly marked
as alpha and unstable in the docs.