Explicitly require TAP::Stream 0.44
Some uses of the toolchain (like our internal minicpan manager) will get very
confused without this version number. In our case, when I added a new TCM to
our minicpan, it didn't follow the TAP::Stream dependency for some reason. I
suspect this is because there was already a version in our minicpan as part of
the _old_ TCM.

This explicit dependency should force dependency resolution to see the
separate distro (and it won't hurt anything).