Move more code to the shared Executor role
The Sequential role is basically empty now and the Parallel role is using all
of the sequential code in the Executor role for tests that cannot be run in
parallel.

Also fixed the Parallel Executor to use the right set of subtests. Previously
it only made a per-class subtest if the class had > 1 test instance, but it
_always_ made a per-test instance subtest. This ended up looking identical to
the Sequential output for classes with only one instance, but would be wrong
for classes with > 1 instance.