There's no need for our own custom context_do
This didn't seem to improve the trace at all.

Also added some tests of the Test2::Util::Trace object associated with some
events to make sure it looks sane. The line number reporting could probably be
better, but I'm not sure exactly how to achieve that with the Test2 tools.